requirements.txt
python-dotenv
eikon
fastapi
pandas
uvicorn